Fixed time-map summary formatting. Moved certain back from common.js, due to it being too early to tell if it makes sense for them to be there.
